Add native endpoints for applications management
