+++ lib/commands/index.js
@@ -16,8 +16,9 @@ import fileMovementExtensions from './file-movement';
 import screenshotExtensions from './screenshots';
 import pasteboardExtensions from './pasteboard';
 import locationExtensions from './location';
-import recordscreenExtensions from './recordscreen';
+import recordScreenExtensions from './recordscreen';
 import lockExtensions from './lock';
+import appManagementExtensions from './app-management';
 
 let commands = {};
 
@@ -26,7 +27,7 @@ Object.assign(commands, actionsCommands, contextCommands, executeExtensions,
   generalExtensions, logExtensions, webExtensions, timeoutExtensions,
   navigationExtensions, elementExtensions, fileMovementExtensions,
   alertExtensions, screenshotExtensions, pasteboardExtensions, locationExtensions,
-  recordscreenExtensions, lockExtensions
+  lockExtensions, recordScreenExtensions, appManagementExtensions
 );
 
 export default commands;
