+++ package.json
@@ -27,7 +27,7 @@
     "lib": "lib"
   },
   "dependencies": {
-    "appium-base-driver": "^2.14.0",
+    "appium-base-driver": "^2.20.3",
     "appium-ios-driver": "^1.30.0",
     "appium-ios-simulator": "^2.6.0",
     "appium-support": "^2.13.0",
